Change types to int

+++ modules/task_3/gushchin_a_radix_with_merge/radix_with_merge.h
@@ -25,16 +25,16 @@ std::vector<T> generateRandomVector(const int size) {
 }
 
 template<typename T>
-std::vector<long> calculateCounters(const std::vector<T>& data) {
+std::vector<int> calculateCounters(const std::vector<T>& data) {
     auto dataSize = data.size();
 
-    std::vector<long> counters(256 * sizeof(T));
+    std::vector<int> counters(256 * sizeof(T));
 
     const uchar* dataBytes = reinterpret_cast<const uchar*>(data.data());
     const uchar* dataEnd = reinterpret_cast<const uchar*>(data.data() + dataSize);
 
     while (dataBytes != dataEnd) {
-        for (short i = 0; i < sizeof(T); i++) {
+        for (int i = 0; i < sizeof(T); i++) {
             counters[256 * static_cast<size_t>(i) + *dataBytes]++;
             dataBytes++;
         }
@@ -44,10 +44,10 @@ std::vector<long> calculateCounters(const std::vector<T>& data) {
 }
 
 template<typename T>
-std::vector<T> radixPass(const std::vector<T>& source, std::vector<long> counter, short byteOffset) {
-    long partialSum = 0;
-    long currentCount;
-    for (short i = 0; i < 256; i++) {
+std::vector<T> radixPass(const std::vector<T>& source, std::vector<int> counter, int byteOffset) {
+    int partialSum = 0;
+    int currentCount;
+    for (int i = 0; i < 256; i++) {
         currentCount = counter[i];
         counter[i] = partialSum;
         partialSum += currentCount;
@@ -64,21 +64,21 @@ std::vector<T> radixPass(const std::vector<T>& source, std::vector<long> counter
 }
 
 template<typename T>
-std::vector<T> radixLastPassSigned(const std::vector<T>& source, std::vector<long> counter, short byteOffset) {
-    long numNeg = 0;
-    for (short i = 128; i < 256; i++)
+std::vector<T> radixLastPassSigned(const std::vector<T>& source, std::vector<int> counter, int byteOffset) {
+    int numNeg = 0;
+    for (int i = 128; i < 256; i++)
         numNeg += counter[i];
 
-    long partialSum = numNeg;
-    long currentCount;
-    for (short i = 0; i < 128; i++) {
+    int partialSum = numNeg;
+    int currentCount;
+    for (int i = 0; i < 128; i++) {
         currentCount = counter[i];
         counter[i] = partialSum;
         partialSum += currentCount;
     }
 
     partialSum = 0;
-    for (short i = 128; i < 256; i++) {
+    for (int i = 128; i < 256; i++) {
         currentCount = counter[i];
         counter[i] = partialSum;
         partialSum += currentCount;
@@ -102,7 +102,7 @@ std::vector<T> radixSortSigned(const std::vector<T>& in) {
     auto counters = calculateCounters(in);
 
     for (uchar i = 0; i < sizeof(T); i++) {
-        std::vector<long> counter(counters.begin() + 256 * static_cast<size_t>(i),
+        std::vector<int> counter(counters.begin() + 256 * static_cast<size_t>(i),
                                   counters.begin() + 256 * static_cast<size_t>(i) + 256);
 
         if (counter[0] == size)
